Module ledger_canister::account_identifier

Structs

AccountIdentifier
While this is backed by an array of length 28, it's canonical representation is a hex string of length 64. The first 8 characters are the CRC-32 encoded hash of the following 56 characters of hex. Both, upper and lower case characters are valid in the input string and can even be mixed.

Subaccount
Subaccounts are arbitrary 32-byte values.

Statics

SUB_ACCOUNT_ZERO