Module ledger_canister::protobuf

Modules

block_response
Nested message and enum types in BlockResponse.

cycles_notification_response
Nested message and enum types in CyclesNotificationResponse.

get_blocks_response
Nested message and enum types in GetBlocksResponse.

transaction
Nested message and enum types in Transaction.

Structs

Account

AccountBalanceRequest
Get the balance of an account

AccountBalanceResponse

AccountIdentifier

ArchiveAddRequest
Add blocks to the archive canister

ArchiveAddResponse

ArchiveIndexEntry

ArchiveIndexResponse

ArchiveInit

Archive canister * Init the archive canister

Block
This is the

BlockHeight

BlockRequest
Get a single block

BlockResponse

Burn

Certification

CyclesNotificationResponse

EncodedBlock

EncodedBlocks

GetBlocksRequest
Get a set of blocks

GetBlocksResponse

GetNodesRequest
Fetch a list of all of the archive nodes

GetNodesResponse

Hash

IcpTs
** BASIC TYPES **

IterBlocksRequest
Iterate through blocks

IterBlocksResponse

LedgerArchiveRequest
Archive any blocks older than this

LedgerInit
Initialise the ledger canister

LedgerUpgrade
The format of values serialized to/from the stable memory during and upgrade

Memo

Mint

NotifyRequest
Notify a canister that it has recieved a payment

NotifyResponse

Payment

Refund

Send

SendRequest
Make a payment

SendResponse

Subaccount

TimeStamp

TipOfChainRequest
Get the length of the chain with a certification

TipOfChainResponse

ToppedUp

TotalSupplyRequest
How many ICPTs are there not in the minting account

TotalSupplyResponse

Transaction

TransactionNotificationRequest

TransactionNotificationResponse